Use clean exit code when --delay is used and no deps are too outdated
The README file states that no error is thrown if all dependencies are less out
of date than the number of days specified in `--delay`, but an exit code of 1
was still used.

An exit code of 1 is the same as throwing an error so it should instead cause
piprot to use an exit code of 0.

I have at the same time fixed some typos in the word "dependencies" around that
part of the code and reflected it in the example output in the README file too.